+++ backend/src/main/java/org/donggle/backend/application/service/AuthService.java
@@ -5,8 +5,8 @@ import org.donggle.backend.application.repository.CategoryRepository;
 import org.donggle.backend.application.repository.MemberCredentialsRepository;
 import org.donggle.backend.application.repository.MemberRepository;
 import org.donggle.backend.application.repository.TokenRepository;
-import org.donggle.backend.application.service.oauth.kakao.LoginClients;
-import org.donggle.backend.application.service.oauth.kakao.SocialType;
+import org.donggle.backend.application.service.oauth.LoginClients;
+import org.donggle.backend.application.service.oauth.SocialType;
 import org.donggle.backend.application.service.oauth.kakao.dto.UserInfo;
 import org.donggle.backend.application.service.request.OAuthAccessTokenRequest;
 import org.donggle.backend.auth.JwtTokenProvider;
