+++ backend/src/main/java/org/donggle/backend/application/service/oauth/LoginClient.java
@@ -1,4 +1,4 @@
-package org.donggle.backend.application.service.oauth.kakao;
+package org.donggle.backend.application.service.oauth;
 
 import org.donggle.backend.application.service.oauth.kakao.dto.UserInfo;
