+++ backend/src/main/java/org/donggle/backend/application/service/oauth/kakao/dto/KakaoProfileResponse.java
@@ -1,7 +1,7 @@
 package org.donggle.backend.application.service.oauth.kakao.dto;
 
 import com.fasterxml.jackson.annotation.JsonProperty;
-import org.donggle.backend.application.service.oauth.kakao.SocialType;
+import org.donggle.backend.application.service.oauth.SocialType;
 
 import static org.donggle.backend.application.service.oauth.kakao.dto.KakaoProfileResponse.KakaoAccount.Profile;
